Moved the submenus and DefaultConfig into their own directory.
